vexpress64: fvp: Fix size of DRAM available to U-boot
On both the Base FVPs and Juno development platform, the ARM Trusted
Firmware configures the TrustZone Controller (TZC) to create a region
in the DRAM which is accessible only from the secure state. This
region corresponds to the top 16MB of DRAM.

Although the Juno port correctly accounts for this secure region,
the FVP port doesn't. This patch fixes this issue. It also removes
the distinction between the 2MB used for SCP DDR retraining and the
remaining 14MB of the secure region on Juno, as U-boot doesn't need
to know that and should just consider it all as unaccessible memory.

Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>