HACK: tc2 baseaddr 0xb0000000
Change the base address of U-Boot to match UEFI so that we don't have to
update SITE1/HBI0249A/images.txt when we swap between U-Boot and UEFI.

This patch is marked as HACK because it will probably never be
upstreamed.  It's also possible that the UEFI base address is changed
instead.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>
Signed-off-by: Asha R <asha.r@arm.com>