Add vexpress_juno_aarch32
Add Aarch32 variant of Juno.

TODO:

CONFIG_BOOTCOMMAND between Juno and FVP could actually be shared.

Currently, Juno uses "afs load" and FVP uses "smhload". This could be
moved into a #define.

The script itself is attempting to do basically the same thing. The main
difference is that Juno has added an alternate kernel and fdt filename.
We could use this same mechanism on FVP also.

So the only difference between FVP and Juno in vexpress_aemv8a.h should
be the CONFIG_BOOTARGS, because FVP sets androidboot.hardware to
"fvpbase", where Juno sets it to "juno". Again, this could be added to a
define in the header or in the defconfig.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>
Signed-off-by: Asha R <asha.r@arm.com>