vexpress64: fvp: boot from DRAM if semihosting fails
If devtree.dtb, ramdisk.img and Image exist, boot via semihosting.

If we partially load from semihosting, then fail, stop loading.

If nothing loads from semihosting, boot from whatever is in DRAM.

note: semihosting expects to load an initrd without a u-boot header and
sets the chosen node with the initrd address and length. We don't know
the length of the initrd in the DRAM method, so it will need a u-boot
header so booti can load it.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>
Signed-off-by: Asha R <asha.r@arm.com>