arm: imx: hab: Add IVT header definitions
The i.MX BootROMs containing the High Assurance Boot (HAB) block relies on
a data structure called the Image Vector Table (IVT) to describe to the
BootROM HAB verifications routines where to locate various data-structures
used by HAB during authentication.

This patch adds the basics of the IVT header descriptions for use in later
patches to parse and validate the IVT to a greater extent than is currently
done.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>