warp7: hab: Add routine to discover HAB load address
This patch takes a given address for example the load address of a zimage
and translates it into the IVT load address based on the zImage load
address.

Since the IVT and HAB used fixed addresses instead of relative offsets to a
base address it is necessary to load HAB images to specific addresses. The
address is always relative to the load entry-point of the binary image,
which is passed to u-boot's mkimage when generating an IVT prefixed binary.

We should only have to specify the load address of a binary once in
non-secure mode, the u-boot BSP port for WaRP7 then will derive the correct
IVT address based on the non-IVT prefixed non-secure load address.

To that calculation neatly and at a high-level this patch makes a u-boot
command which does the necessary offset subtraction and stores the result
in an environment variable.

Since the IVT/CSF pair use fixed offsets but aren't bothered about where
they live relative to each other - or indeed to the binary they sign we
can't put a HAB command into hab.c that assumes IVT | BINARY | CSF - since
the NXP secure boot standard doesn't mandate that.

The best we can do is say for WaRP7 secure boot we support this one option
only.

Hence the function for making the calculation is associated with WaARP7 as
opposed to being buried in hab.c.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>