arm: imx: hab: Make authenticate_image return int
Both usages of authenticate_image treat the result code as a simple binary,
moreover the version bound to the console callback wants the result code to
be a signed int not a uint32_t. Later patches will fixup the internal
return value, for now this patch fixes the return type only.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>