arm: imx: hab: Fix authenticate_image result code
authenticate_image returns 1 for success and 0 for failure. That result
code is mapped directly to the result code for the command line function
hab_auth_img - which means when hab_auth_img succeeds it is returning
CMD_RET_FAILURE (1) instead of CMD_RET_SUCCESS (0).

This patch fixes this by making authenticate_image() return 0 for success
and 1 for failure. Both users of authenticate_image() as a result have some
minimal churn.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>