warp7: hab: Add warp7_failsafe to the HAB authentication failure path
If authentication via HAB should fail then the next thing the bootloader
should do is drop into a secure recovery mode - in this case the HAB API
callback failsafe().

Running the failsafe command when the board is in non-secure mode will
result in any bootable image being run or if the board is in secure mode
will result in only signed images being run.

For WaRP7 the way the boot scripts are currently constructed only the
secure-boot path (when the secure boot bit is set) will run hab_failure via
the config command warp7_failsafe.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>