warp7: Make boot command a Kconfig selection
For secure boot we probably want to move to having bootloader scr, kernel,
DTB and OPTEE binaries stored on ext4. This change makes it possible to
have FAT or ext4 as the rootfs medium. A subsequent patch to the secure and
non-secure config files will select FAT for both with future patches
selecting ext4 for the secure version once we are producing our secure-boot
image in that format.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Reviewed-by: Rui Miguel Silva <rui.silva@linaro.org>