warp7: Add filesystem path prefixes for ext4 loads
On WaARP7 for secure-optee boots we want to locate our binaries inside of
an ext4 filesystem in standard paths, as opposed to having a dedicated
partition with all of our binaries aggregated in one directory.

This patch sets a new path for OPTEE, DTB, Kernel and boot.scr for
secure-optee boot, when booting with CONFIG_LOADCMD_EXT4.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Reviewed-by: Rui Miguel Silva <rui.silva@linaro.org>