warp7: Make UUID selection contingent on fs layout
The ext4 filesystem layout uses a different UUID for the root device than
the default WaRP7 image produced by NXP's yocoto build.

Change the ext4 build UUID device from 0:2 to 0:1. With the uuidpart made
into an environment variable it will be possible to tell the kernel which
filesystem to mount as the root partition by way of boot.scr.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Reviewed-by: Rui Miguel Silva <rui.silva@linaro.org>