fit: reuse sunxi ARM64 fit src file to ease yocoto build.
ATF sub image can be added back when ATF is available for
ARM32.

The signature of sub-image is needed so that SPL can validate
these images. The signature of configuration is needed, so that
"required" property in signature field of uboot dtb is set to
conf, then kernel/dtb/optee FIT image can be verified as a
combination(FIT configuration).

Signed-off-by: Jun Nie <jun.nie@linaro.org>