imx: mx7: avoid some initialization if low level is skipped
We can have the case where u-boot is launched after some other low level
enabler, like for example when u-boot runs after arm-trusted-firmware and/or
optee. So, because of that we may need to jump the initialization of some IP
blocks even because we may no longer have the permission for that.

So, if the config option to skip low level init is set disable also timer, board
and cup initialization.

Signed-off-by: Rui Miguel Silva <rui.silva@linaro.org>