vexpress: Add SP804 timer support.
sp804 timer is a SoC peripheral present in vexpress model.
This timer can be used in u-boot when arch timer is not available
in core for e.g. in Cortex-A5.

armv7: rename sp804 timer to mmio timer

Rename the wrongly init named mmio timer and fix the use of it on a
Cortex-A5 system that does not have arch_timer.

Signed-off-by: Usama Arif <usama.arif@arm.com>
Signed-off-by: Rui Miguel Silva <rui.silva@linaro.org>