apalis-imx8: configure automatic booting on recovery
When booting U-Boot in recovery mode the default boot command for
automatic boots is substituted by bootcmd_mfg. For tezi we want this
command to source the boot script inside the boot container that the uuu
utility loads. The precise load address of this script is the one used
in the mkimage_imx8 utility to generate the boot container.

NOTE: this change affects both tezi U-Boot as well as BSP U-Boot. Since
currently only tezi is supporting recovery mode, this shouldn't be a
problem. Otherwise, consider adding preprocessor safeguards.

Signed-off-by: Gerard Salvatella <gerard.salvatella@toradex.com>