[FIO toup] fpga: imx_m4: boot M4
This driver allows a controlled boot of the M4 processor.
When secure mode is enabled (FIOVB), the hash of the firmware in the
FIT image is compared against the value stored in the mmc RPMB
partition (via OPTEE). If they differ, the firmware in QSPI flash will
be upgraded to the new image before booting.

When secure mode is not enabled (!FIOVB), the hash of the firmware in
the FIT image is compared against the value calculated from QSPI flash
(but using the size of the FIT image firmware as the size of the QSPI
image in flash). If they difer, the firmware in QSPI flash will be
upgraded to the new image before booting.

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>

- fix compilation errors
- fix infinite loop on corrupted flash data

If the flash image has been erased or it cant be booted, the data
being help in RPMB is no longer valid to compare against potential
firmwares in the itb.

- correct check for M4 running
- handle dual_boot (M4/A7 running)

In dual boot mode, the M4 needs to be restarted after an upgrade -
this is a board specific operation.

Upstream-Status: Pending
Signed-off-by: Michael Scott <mike@foundries.io>
Signed-off-by: Jorge Ramirez <jorge@foundries.io>
Co-developed-by: Oleksandr Suvorov <oleksandr.suvorov@foundries.io>
Signed-off-by: Oleksandr Suvorov <oleksandr.suvorov@foundries.io>