zynqmp: restore the jtag interface
As a security feature, if boot.bin was configured for secure boot the
CSU will disable the JTAG interface on all cases.

Some boards might rely on this interface for flashing to QSPI in which
case those systems might end up bricked during development

This commit attempts to restore the interface - if the CSU allows for it.

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>