[FIO fromlist] tools: image-host: print error messages to stderr
The make by default cuts off the stdout output from external tools,
so all error messages from the image-host are not shown in a make
output. Besides that, it is a common approach to use stderr stream
for error messages.
Use stderr for all error messages in image-host.

Upstream-Status: Accepted [v2024.01]
Signed-off-by: Oleksandr Suvorov <oleksandr.suvorov@foundries.io>